src/etl/requirements.txt
boto3==1.35.97
pandas==2.2.3
playwright==1.41.0
PyMySQL==1.1.1
python-dotenv==1.0.1